+++ database/migrations/2020_01_06_201832_create_kenshin_sys_options_table.php
@@ -18,7 +18,7 @@ class CreateKenshinSysOptionsTable extends Migration
         Schema::create('kenshin_sys_options', function (Blueprint $table) {
             $table->increments('id');
             $table->bigInteger('kenshin_sys_course_id')->unsigned();
-            $table->integer('kenshin_sys_option_no');
+            $table->bigInteger('kenshin_sys_option_no');
             $table->string('kenshin_sys_option_name', 100);
             $table->integer('kenshin_sys_option_age_kisan_kbn');
             $table->date('kenshin_sys_option_age_kisan_date')->nullable();
